Move roll logic out of vue component
